recommit

+++ FinalProject/src/main/java/com/team12/main/team1/shop/review/Team1ReviewMapper.java
@@ -21,7 +21,7 @@ public interface Team1ReviewMapper {
 	int getAllReview(Team1ReviewDTO review);
 	
 	// 전체 별점 평균 구하기
-	int getAllReviewAvg();
+	double getAllReviewAvg();
